Week_03/requirements.txt
faiss-cpu
sentence-transformers
numpy
fastapi
uvicorn
python-dotenv
groq
